Fe/students

+++ src/MathInsight.Modules.Grading_Analytics/Persistence/Entities/TagDifficultyReadOnly.cs
@@ -7,5 +7,5 @@ namespace MathInsight.Modules.Grading_Analytics.Persistence.Entities;
 public class TagDifficultyReadOnly
 {
     public string DifficultyId { get; set; } = string.Empty;
-    public byte LevelValue { get; set; }
+    public int LevelValue { get; set; }
 }
